+++ saleor/graphql/core/mutations.py
@@ -1,5 +1,5 @@
 from itertools import chain
-from typing import Tuple
+from typing import Tuple, Union
 
 import graphene
 from django.contrib.auth import get_user_model
@@ -9,6 +9,7 @@ from django.core.exceptions import (
     ValidationError,
 )
 from django.db.models.fields.files import FileField
+from graphene import ObjectType
 from graphene.types.mutation import MutationOptions
 from graphene_django.registry import get_global_registry
 from graphql.error import GraphQLError
@@ -19,7 +20,7 @@ from ...account import models
 from ..account.types import User
 from ..utils import get_nodes
 from .types import Error, MetaInput, MetaPath, Upload
-from .utils import snake_to_camel_case
+from .utils import from_global_id_strict_type, snake_to_camel_case
 
 registry = get_global_registry()
 
@@ -104,12 +105,36 @@ class BaseMutation(graphene.Mutation):
         cls._meta.fields.update(fields)
 
     @classmethod
-    def get_node_or_error(cls, info, node_id, field="id", only_type=None):
+    def get_node_by_pk(
+        cls, info, graphene_type: ObjectType, pk: Union[int, str], qs=None
+    ):
+        """Attempt to resolve a node from the given internal ID.
+
+        Whether by using the provided query set object or by calling type's get_node().
+        """
+        if qs is not None:
+            return qs.filter(pk=pk).first()
+        get_node = getattr(graphene_type, "get_node", None)
+        if get_node:
+            return get_node(info, pk)
+        return None
+
+    @classmethod
+    def get_node_or_error(cls, info, node_id, field="id", only_type=None, qs=None):
         if not node_id:
             return None
 
         try:
-            node = graphene.Node.get_node_from_global_id(info, node_id, only_type)
+            if only_type is not None:
+                pk = from_global_id_strict_type(node_id, only_type, field=field)
+            else:
+                # FIXME: warn when supplied only_type is None?
+                only_type, pk = graphene.Node.from_global_id(node_id)
+
+            if isinstance(only_type, str):
+                only_type = info.schema.get_type(only_type).graphene_type
+
+            node = cls.get_node_by_pk(info, graphene_type=only_type, pk=pk, qs=qs)
         except (AssertionError, GraphQLError) as e:
             raise ValidationError({field: str(e)})
         else:
@@ -331,6 +356,10 @@ class ModelMutation(BaseMutation):
 
     @classmethod
     def get_instance(cls, info, **data):
+        """Retrieve an instance from the supplied global id.
+
+        The expected graphene type can be lazy (str).
+        """
         object_id = data.get("id")
         if object_id:
             model_type = registry.get_type_for_model(cls._meta.model)