+++ tests/conftest.py
@@ -11,7 +11,6 @@ from django.core.files import File
 from django.core.files.uploadedfile import SimpleUploadedFile
 from django.forms import ModelForm
 from django.test.client import Client
-from django.utils.encoding import smart_text
 from django_countries import countries
 from PIL import Image
 from prices import Money, TaxedMoney
@@ -48,6 +47,7 @@ from saleor.product.models import (
     ProductType,
     ProductVariant,
 )
+from saleor.product.utils.attributes import associate_attribute_values_to_instance
 from saleor.shipping.models import ShippingMethod, ShippingMethodType, ShippingZone
 from saleor.site import AuthenticationBackends
 from saleor.site.models import AuthorizationKey, SiteSettings
@@ -388,16 +388,15 @@ def categories_tree(db, product_type):  # pylint: disable=W0613
 
     product_attr = product_type.product_attributes.first()
     attr_value = product_attr.values.first()
-    attributes = {smart_text(product_attr.pk): [smart_text(attr_value.pk)]}
 
-    Product.objects.create(
+    product = Product.objects.create(
         name="Test product",
-        price=Money("10.00", "USD"),
+        price=Money(10, "USD"),
         product_type=product_type,
-        attributes=attributes,
         category=child,
     )
 
+    associate_attribute_values_to_instance(product, product_attr, attr_value)
     return parent
 
 
@@ -452,31 +451,29 @@ def product_type_without_variant():
 @pytest.fixture
 def product(product_type, category):
     product_attr = product_type.product_attributes.first()
-    attr_value = product_attr.values.first()
-    attributes = {smart_text(product_attr.pk): [smart_text(attr_value.pk)]}
+    product_attr_value = product_attr.values.first()
 
     product = Product.objects.create(
         name="Test product",
         price=Money("10.00", "USD"),
         product_type=product_type,
-        attributes=attributes,
         category=category,
     )
 
+    associate_attribute_values_to_instance(product, product_attr, product_attr_value)
+
     variant_attr = product_type.variant_attributes.first()
     variant_attr_value = variant_attr.values.first()
-    variant_attributes = {
-        smart_text(variant_attr.pk): [smart_text(variant_attr_value.pk)]
-    }
 
-    ProductVariant.objects.create(
+    variant = ProductVariant.objects.create(
         product=product,
         sku="123",
-        attributes=variant_attributes,
         cost_price=Money("1.00", "USD"),
         quantity=10,
         quantity_allocated=1,
     )
+
+    associate_attribute_values_to_instance(variant, variant_attr, variant_attr_value)
     return product
 
 
@@ -497,10 +494,7 @@ def product_with_multiple_values_attributes(product, product_type, category) -> 
     product_type.product_attributes.clear()
     product_type.product_attributes.add(attribute)
 
-    product.attributes = {
-        smart_text(attribute.pk): [smart_text(attr_val_1.pk), smart_text(attr_val_2.pk)]
-    }
-    product.save(update_fields=["attributes"])
+    associate_attribute_values_to_instance(product, attribute, attr_val_1, attr_val_2)
     return product
 
 
@@ -508,7 +502,7 @@ def product_with_multiple_values_attributes(product, product_type, category) -> 
 def product_with_default_variant(product_type_without_variant, category):
     product = Product.objects.create(
         name="Test product",
-        price=Money("10.00", "USD"),
+        price=Money(10, "USD"),
         product_type=product_type_without_variant,
         category=category,
     )
@@ -562,38 +556,36 @@ def product_without_shipping(category):
 def product_list(product_type, category):
     product_attr = product_type.product_attributes.first()
     attr_value = product_attr.values.first()
-    attributes = {smart_text(product_attr.pk): [smart_text(attr_value.pk)]}
 
-    products = Product.objects.bulk_create(
-        [
-            Product(
-                pk=1486,
-                name="Test product 1",
-                price=Money("10.00", "USD"),
-                category=category,
-                product_type=product_type,
-                attributes=attributes,
-                is_published=True,
-            ),
-            Product(
-                pk=1487,
-                name="Test product 2",
-                price=Money("20.00", "USD"),
-                category=category,
-                product_type=product_type,
-                attributes=attributes,
-                is_published=False,
-            ),
-            Product(
-                pk=1489,
-                name="Test product 3",
-                price=Money("20.00", "USD"),
-                category=category,
-                product_type=product_type,
-                attributes=attributes,
-                is_published=True,
-            ),
-        ]
+    products = list(
+        Product.objects.bulk_create(
+            [
+                Product(
+                    pk=1486,
+                    name="Test product 1",
+                    price=Money(10, "USD"),
+                    category=category,
+                    product_type=product_type,
+                    is_published=True,
+                ),
+                Product(
+                    pk=1487,
+                    name="Test product 2",
+                    price=Money(20, "USD"),
+                    category=category,
+                    product_type=product_type,
+                    is_published=False,
+                ),
+                Product(
+                    pk=1489,
+                    name="Test product 3",
+                    price=Money(20, "USD"),
+                    category=category,
+                    product_type=product_type,
+                    is_published=True,
+                ),
+            ]
+        )
     )
     ProductVariant.objects.bulk_create(
         [
@@ -617,6 +609,10 @@ def product_list(product_type, category):
             ),
         ]
     )
+
+    for product in products:
+        associate_attribute_values_to_instance(product, product_attr, attr_value)
+
     return products
 
 
@@ -679,19 +675,16 @@ def unavailable_product_with_variant(product_type, category):
 
     variant_attr = product_type.variant_attributes.first()
     variant_attr_value = variant_attr.values.first()
-    variant_attributes = {
-        smart_text(variant_attr.pk): [smart_text(variant_attr_value.pk)]
-    }
 
-    ProductVariant.objects.create(
+    variant = ProductVariant.objects.create(
         product=product,
         sku="123",
-        attributes=variant_attributes,
-        cost_price=Money("1.00", "USD"),
+        cost_price=Money(1, "USD"),
         quantity=10,
         quantity_allocated=1,
     )
 
+    associate_attribute_values_to_instance(variant, variant_attr, variant_attr_value)
     return product