+++ tests/test_product_attributes.py
@@ -3,76 +3,75 @@ from unittest.mock import MagicMock, Mock
 import pytest
 from prices import Money
 
-from saleor.product.models import Attribute, AttributeValue, Product, ProductType
+from saleor.product import AttributeInputType
+from saleor.product.models import AttributeValue, Product, ProductType, ProductVariant
 from saleor.product.tasks import _update_variants_names
 from saleor.product.utils.attributes import (
-    generate_name_from_values,
-    get_attributes_display_map,
-    get_name_from_attributes,
+    associate_attribute_values_to_instance,
+    generate_name_for_variant,
 )
 
 
 @pytest.fixture()
-def product_with_no_attributes(product_type, category):
+def variant_with_no_attributes(category):
+    """Create a variant having no attributes, the same for the parent product."""
+    product_type = ProductType.objects.create(
+        name="Test product type", has_variants=True, is_shipping_required=True
+    )
     product = Product.objects.create(
         name="Test product",
         price=Money(10, "USD"),
         product_type=product_type,
         category=category,
     )
-    return product
+    variant = ProductVariant.objects.create(product=product, sku="123")
+    return variant
 
 
-def test_get_attributes_display_map(product):
-    attributes = product.product_type.product_attributes.all()
-    attributes_display_map = get_attributes_display_map(product, attributes)
+def test_generate_name_for_variant(
+    variant_with_no_attributes, color_attribute_without_values, size_attribute
+):
+    """Test the name generation from a given variant containing multiple attributes and
+    different input types (dropdown and multiselect).
+    """
 
-    product_attr = product.product_type.product_attributes.first()
-    attr_value = product_attr.values.first()
+    variant = variant_with_no_attributes
+    color_attribute = color_attribute_without_values
 
-    assert len(attributes_display_map) == 1
-    assert {k: v.pk for k, v in attributes_display_map.items()} == {
-        product_attr.pk: attr_value.translated.pk
-    }
-
-
-def test_get_attributes_display_map_empty(product_with_no_attributes):
-    product = product_with_no_attributes
-    attributes = product.product_type.product_attributes.all()
-    assert get_attributes_display_map(product, attributes) == {}
-
-
-def test_get_name_from_attributes(product):
-    variant = product.variants.first()
-    attributes = variant.product.product_type.variant_attributes.all()
-    name = get_name_from_attributes(variant, attributes)
-    assert name == "Small"
-
-
-def test_get_name_from_attributes_no_attributes(product_with_no_attributes):
-    variant_without_attributes = product_with_no_attributes.variants.create(
-        sku="example-sku"
+    # Assign the attributes to the product type
+    variant.product.product_type.variant_attributes.set(
+        (color_attribute, size_attribute)
     )
-    variant = variant_without_attributes
-    attributes = variant.product.product_type.variant_attributes.all()
-    name = get_name_from_attributes(variant, attributes)
-    assert name == ""
 
+    # Set the color attribute to a multi-value attribute
+    color_attribute.input_type = AttributeInputType.MULTISELECT
+    color_attribute.save(update_fields=["input_type"])
 
-def test_generate_name_from_values():
-    attribute = Attribute.objects.create(slug="color", name="Color")
-    red = AttributeValue.objects.create(attribute=attribute, name="Red", slug="red")
-    blue = AttributeValue.objects.create(attribute=attribute, name="Blue", slug="blue")
-    yellow = AttributeValue.objects.create(
-        attribute=attribute, name="Yellow", slug="yellow"
+    # Create colors
+    colors = AttributeValue.objects.bulk_create(
+        [
+            AttributeValue(attribute=color_attribute, name="Yellow", slug="yellow"),
+            AttributeValue(attribute=color_attribute, name="Blue", slug="blue"),
+            AttributeValue(attribute=color_attribute, name="Red", slug="red"),
+        ]
     )
-    values = {"3": red, "2": blue, "1": yellow}
-    name = generate_name_from_values(values)
-    assert name == "Yellow / Blue / Red"
+
+    # Retrieve the size attribute value "Big"
+    size = size_attribute.values.get(slug="big")
+
+    # Associate the colors and size to variant attributes
+    associate_attribute_values_to_instance(variant, color_attribute, *tuple(colors))
+    associate_attribute_values_to_instance(variant, size_attribute, size)
+
+    # Generate the variant name from the attributes
+    name = generate_name_for_variant(variant)
+    assert name == "Yellow, Blue, Red / Big"
 
 
-def test_generate_name_from_values_empty():
-    name = generate_name_from_values({})
+def test_generate_name_from_values_empty(variant_with_no_attributes):
+    """Ensure generate a variant name from a variant without any attributes assigned
+    returns an empty string."""
+    name = generate_name_for_variant(variant_with_no_attributes)
     assert name == ""
 
 
@@ -94,3 +93,46 @@ def test_update_variants_changed_does_nothing_with_no_attributes():
     product_type.variant_attributes.all = Mock(return_value=[])
     saved_attributes = []
     assert _update_variants_names(product_type, saved_attributes) is None
+
+
+def test_associate_attribute_to_non_product_instance(color_attribute):
+    instance = ProductType()
+    attribute = color_attribute
+    value = color_attribute.values.first()
+
+    with pytest.raises(AssertionError) as exc:
+        associate_attribute_values_to_instance(instance, attribute, value)  # noqa
+
+    assert exc.value.args == ("ProductType is unsupported",)
+
+
+def test_associate_attribute_to_product_instance_from_different_attribute(
+    product, color_attribute, size_attribute
+):
+    """Ensure an assertion error is raised when one tries to associate attribute values
+    to an object that don't belong to the supplied attribute.
+    """
+    instance = product
+    attribute = color_attribute
+    value = size_attribute.values.first()
+
+    with pytest.raises(AssertionError) as exc:
+        associate_attribute_values_to_instance(instance, attribute, value)
+
+    assert exc.value.args == ("Some values are not from the provided attribute.",)
+
+
+def test_associate_attribute_to_product_instance_without_values(product):
+    """Ensure clearing the values from a product is properly working."""
+    old_assignment = product.attributes.first()
+    assert old_assignment is not None, "The product doesn't have attribute-values"
+    assert old_assignment.values.count() == 1
+
+    attribute = old_assignment.attribute
+
+    # Clear the values
+    new_assignment = associate_attribute_values_to_instance(product, attribute)
+
+    # Ensure the values were cleared and no new assignment entry was created
+    assert new_assignment.pk == old_assignment.pk
+    assert new_assignment.values.count() == 0