+++ CHANGELOG.md
@@ -66,6 +66,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Fix for Digital products - update order status after automatic fulfillment - #4709 by @korycins
 - Add redirectUrl to staff and user create mutations - #4717 by @fowczarek
 - Filtering: use search in a consistent way. Add sort by product type name and publication status to products query. - #4715 by @fowczarek
+- Migrated the old product attributes mapping to M2M - #4663 by @NyanKiyoshi
 
 ## 2.8.0